Add files via upload
This is the first version of the node.js web server that should create a local server on your machine. It worked for me and here's a link with my step by step tutorial on how it worked. I just wanted to put this out there so everyone can begin working at least on a local server on their machine. https://www.digitalocean.com/community/tutorials/how-to-create-a-web-server-in-node-js-with-the-http-module